Fix file location issues
