+++ app-modules/report/src/Filament/Pages/StudentEngagementReport.php
@@ -36,12 +36,12 @@
 
 namespace AdvisingApp\Report\Filament\Pages;
 
-use Filament\Pages\Dashboard;
-use App\Filament\Clusters\ReportLibrary;
+use AdvisingApp\Report\Filament\Widgets\MostEngagedStudentsTable;
 use AdvisingApp\Report\Filament\Widgets\RefreshWidget;
-use AdvisingApp\StudentDataModel\Filament\Widgets\StudentEngagementStats;
-use AdvisingApp\StudentDataModel\Filament\Widgets\MostEngagedStudentsTable;
-use AdvisingApp\StudentDataModel\Filament\Widgets\StudentEngagementLineChart;
+use AdvisingApp\Report\Filament\Widgets\StudentEngagementLineChart;
+use AdvisingApp\Report\Filament\Widgets\StudentEngagementStats;
+use App\Filament\Clusters\ReportLibrary;
+use Filament\Pages\Dashboard;
 
 class StudentEngagementReport extends Dashboard
 {
