+++ app-modules/report/src/Filament/Widgets/StudentEngagementLineChart.php
@@ -34,7 +34,7 @@
 </COPYRIGHT>
 */
 
-namespace AdvisingApp\StudentDataModel\Filament\Widgets;
+namespace AdvisingApp\Report\Filament\Widgets;
 
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Cache;
